move tools into src/tools
