+++ src/tools/CleanData/index.tsx
@@ -1,5 +1,5 @@
 import { css } from "styled-components"
-import { Textarea } from "../Textarea"
+import { Textarea } from "../../components/Textarea"
 import { useInput } from "../../hooks/useInput"
 import { cleanData } from "./cleanData"
 
@@ -8,15 +8,17 @@ export function CleanData(): JSX.Element {
   const cleaned = cleanData(text)
 
   return (
-    <div css={css`
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      align-items: center;
-      & > * + * {
-        margin-top: 1em;
-      }
-    `}>
+    <div
+      css={css`
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        & > * + * {
+          margin-top: 1em;
+        }
+      `}
+    >
       <Textarea value={text} onChange={change} />
       <Textarea value={cleaned} readOnly />
     </div>
